backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pgvector==0.2.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Embeddings and RAG - FIXED FOR PYTHON 3.12
sentence-transformers==2.3.1
numpy==1.26.2

# MCP SDK
anthropic==0.8.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0